+++ src/commands/hardis/org/create.ts
@@ -151,7 +151,7 @@ export default class SandboxCreate extends SfdxCommand {
     await fs.ensureDir(path.dirname(projectSandboxDefLocal));
     await fs.writeFile(projectSandboxDefLocal, JSON.stringify(this.projectSandboxDef, null, 2));
 
-    // Fix sfdx-cli bug: remove shape.zip if found
+    // Fix @salesforce/cli bug: remove shape.zip if found
     const tmpShapeFolder = path.join(os.tmpdir(), "shape");
     if (fs.existsSync(tmpShapeFolder)) {
       await fs.remove(tmpShapeFolder);
@@ -195,16 +195,16 @@ export default class SandboxCreate extends SfdxCommand {
     } else if (createResult.status === 1 && createResult.errorMessage.includes("Socket timeout occurred while listening for results")) {
       return c.red(
         `[sfdx-hardis] Error creating sandbox org. ${c.bold(
-          "This is probably a Salesforce error, try again manually or launch again CI job"
-        )}\n${JSON.stringify(createResult, null, 2)}`
+          "This is probably a Salesforce error, try again manually or launch again CI job",
+        )}\n${JSON.stringify(createResult, null, 2)}`,
       );
     }
     return c.red(
       `[sfdx-hardis] Error creating sandbox org. Maybe try ${c.yellow(c.bold("sfdx hardis:sandbox:create --forcenew"))} ?\n${JSON.stringify(
         createResult,
         null,
-        2
-      )}`
+        2,
+      )}`,
     );
   }