+++ src/commands/hardis/scratch/pool/view.ts
@@ -59,8 +59,8 @@ export default class ScratchPoolView extends SfdxCommand {
       uxLog(
         this,
         c.yellow(
-          `There is not scratch orgs pool configured on this project. Please see with your tech lead about using command hardis:scratch:pool:configure`
-        )
+          `There is not scratch orgs pool configured on this project. Please see with your tech lead about using command hardis:scratch:pool:configure`,
+        ),
       );
       return { status: 1, outputString: "Scratch org pool configuration to create" };
     }