+++ src/common/gitProvider/azureDevops.ts
@@ -3,7 +3,12 @@ import * as azdev from "azure-devops-node-api";
 import * as c from "chalk";
 import { uxLog } from "../utils";
 import { PullRequestMessageRequest, PullRequestMessageResult } from ".";
-import { CommentThreadStatus, GitPullRequestCommentThread } from "azure-devops-node-api/interfaces/GitInterfaces";
+import {
+  CommentThreadStatus,
+  GitPullRequestCommentThread,
+  PullRequestAsyncStatus,
+  PullRequestStatus,
+} from "azure-devops-node-api/interfaces/GitInterfaces";
 
 export class AzureDevopsProvider extends GitProviderRoot {
   private azureApi: InstanceType<typeof azdev.WebApi>;
@@ -22,6 +27,46 @@ export class AzureDevopsProvider extends GitProviderRoot {
     return "sfdx-hardis Azure Devops connector";
   }
 
+  public async getBranchDeploymentCheckId(gitBranch: string): Promise<string> {
+    let deploymentCheckId = null;
+    // Get Azure Git API
+    const azureGitApi = await this.azureApi.getGitApi();
+    const repositoryId = process.env.BUILD_REPOSITORY_ID || null;
+    if (repositoryId == null) {
+      uxLog(this, c.yellow("BUILD_REPOSITORY_ID must be defined"));
+      return null;
+    }
+    const latestPullRequestsOnBranch = await azureGitApi.getPullRequests(repositoryId, {
+      targetRefName: `refs/heads/${gitBranch}`,
+      status: PullRequestStatus.Completed,
+    });
+    const latestMergedPullRequestOnBranch = latestPullRequestsOnBranch.filter((pr) => pr.mergeStatus === PullRequestAsyncStatus.Succeeded);
+    if (latestMergedPullRequestOnBranch.length > 0) {
+      const latestPullRequest = latestMergedPullRequestOnBranch[0];
+      const latestPullRequestId = latestPullRequest.pullRequestId;
+      const existingThreads = await azureGitApi.getThreads(repositoryId, latestPullRequestId);
+      for (const existingThread of existingThreads) {
+        if (existingThread.isDeleted) {
+          continue;
+        }
+        for (const comment of existingThread?.comments || []) {
+          if ((comment?.content || "").includes(`<!-- sfdx-hardis deployment-id `)) {
+            const matches = /<!-- sfdx-hardis deployment-id (.*) -->/gm.exec(comment.content);
+            if (matches) {
+              deploymentCheckId = matches[1];
+              uxLog(this, c.gray(`Found deployment id ${deploymentCheckId} on PR #${latestPullRequestId} ${latestPullRequest.title}`));
+            }
+            break;
+          }
+        }
+        if (deploymentCheckId) {
+          break;
+        }
+      }
+    }
+    return deploymentCheckId;
+  }
+
   // Posts a note on the merge request
   public async postPullRequestMessage(prMessage: PullRequestMessageRequest): Promise<PullRequestMessageResult> {
     // Get CI variables
@@ -31,6 +76,18 @@ export class AzureDevopsProvider extends GitProviderRoot {
     const pullRequestIdStr = process.env.SYSTEM_PULLREQUEST_PULLREQUESTID || null;
     if (repositoryId == null || pullRequestIdStr == null) {
       uxLog(this, c.grey("[Azure integration] No project and pull request, so no note thread..."));
+      uxLog(
+        this,
+        c.yellow(`Following variables must be defined when available:
+- BUILD_REPOSITORY_ID
+- BUILD_BUILD_ID
+- SYSTEM_JOB_ID
+- SYSTEM_PULLREQUEST_PULLREQUESTID
+- SYSTEM_JOB_DISPLAY_NAME
+- SYSTEM_COLLECTIONURI
+- SYSTEM_TEAMPROJECT
+      `),
+      );
       return { posted: false, providerResult: { info: "No related pull request" } };
     }
     const pullRequestId = Number(pullRequestIdStr);
@@ -62,6 +119,9 @@ _Provided by [sfdx-hardis](https://sfdx-hardis.cloudity.com) from job [${azureJo
     let existingThreadComment: GitPullRequestCommentThread = null;
     let existingThreadCommentId: number = null;
     for (const existingThread of existingThreads) {
+      if (existingThread.isDeleted) {
+        continue;
+      }
       for (const comment of existingThread?.comments || []) {
         if ((comment?.content || "").includes(`<!-- sfdx-hardis message-key ${messageKey} -->`)) {
           existingThreadComment = existingThread;