+++ src/common/gitProvider/index.ts
@@ -1,5 +1,5 @@
 import * as c from "chalk";
-import { isCI, uxLog } from "../utils";
+import { getCurrentGitBranch, isCI, uxLog } from "../utils";
 import { AzureDevopsProvider } from "./azureDevops";
 import { GithubProvider } from "./github";
 import { GitlabProvider } from "./gitlab";
@@ -9,6 +9,18 @@ export abstract class GitProvider {
   static getInstance(): GitProviderRoot {
     // Azure
     if (process.env.SYSTEM_ACCESSTOKEN) {
+      const serverUrl = process.env.SYSTEM_COLLECTIONURI || null;
+      // a Personal Access Token must be defined
+      const token = process.env.CI_SFDX_HARDIS_AZURE_TOKEN || process.env.SYSTEM_ACCESSTOKEN || null;
+      if (serverUrl == null || token == null) {
+        uxLog(
+          this,
+          c.yellow(`To benefit from Azure Pipelines advanced integration, you need to define the following variables as ENV vars:
+- SYSTEM_COLLECTIONURI
+- SYSTEM_ACCESSTOKEN or CI_SFDX_HARDIS_AZURE_TOKEN`),
+        );
+        return null;
+      }
       return new AzureDevopsProvider();
     }
     // Gitlab
@@ -19,7 +31,7 @@ export abstract class GitProvider {
           this,
           c.yellow(`To benefit from Gitlab advanced integration, you need to :
 - Go to Settings -> Access tokens -> create a project token named "SFDX HARDIS BOT" with developer access and scope "api", then copy its value
-- Go to Settings -> CI/CD -> Variables -> Create a masked variable named CI_SFDX_HARDIS_GITLAB_TOKEN, and paste the access token value`)
+- Go to Settings -> CI/CD -> Variables -> Create a masked variable named CI_SFDX_HARDIS_GITLAB_TOKEN, and paste the access token value`),
         );
         return null;
       }
@@ -32,8 +44,8 @@ export abstract class GitProvider {
       uxLog(
         this,
         c.grey(
-          "To use sfdx-hardis GitProvider capabilities, SYSTEM_ACCESSTOKEN, CI_JOB_TOKEN or GITHUB_TOKEN must be accessible for Azure Pipelines, Gitlab or GitHub"
-        )
+          "To use sfdx-hardis GitProvider capabilities, SYSTEM_ACCESSTOKEN, CI_JOB_TOKEN or GITHUB_TOKEN must be accessible for Azure Pipelines, Gitlab or GitHub",
+        ),
       );
     }
     return null;
@@ -41,6 +53,9 @@ export abstract class GitProvider {
 
   static async managePostPullRequestComment(): Promise<void> {
     const gitProvider = GitProvider.getInstance();
+    if (gitProvider == null) {
+      return;
+    }
     const prData = globalThis.pullRequestData;
     const prCommentSent = globalThis.pullRequestCommentSent || false;
     if (prData && gitProvider && prCommentSent === false) {
@@ -61,6 +76,20 @@ export abstract class GitProvider {
       if (postResult && postResult.posted === true) {
         globalThis.pullRequestCommentSent = true;
       }
+    }
+  }
+
+  static async getDeploymentCheckId(): Promise<string> {
+    const gitProvider = GitProvider.getInstance();
+    if (gitProvider == null) {
+      return null;
+    }
+    try {
+      const currentGitBranch = await getCurrentGitBranch();
+      return gitProvider.getBranchDeploymentCheckId(currentGitBranch);
+    } catch (e) {
+      uxLog(this, c.yellow(`Error while trying to retrieve deployment check id:\n${e.message}`));
+      return null;
     }
   }
 }